requirements.txt
-i https://pypi.org/simple
certifi==2023.7.22 ; python_version >= '3.6'
charset-normalizer==3.3.1 ; python_full_version >= '3.7.0'
filelock==3.13.0 ; python_version >= '3.8'
fsspec==2023.10.0 ; python_version >= '3.8'
idna==3.4 ; python_version >= '3.5'
imagehash==4.3.1
jinja2==3.1.2 ; python_version >= '3.7'
joblib==1.3.2 ; python_version >= '3.7'
markupsafe==2.1.3 ; python_version >= '3.7'
matrix-client==0.4.0
mpmath==1.3.0
mypy-extensions==1.0.0
networkx==3.2.1 ; python_version >= '3.9'
numpy==1.26.1
pandas==2.1.2
pillow==10.1.0 ; python_version >= '3.8'
plemmy==0.3.10
python-dateutil==2.8.2 ; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
pytz==2023.3.post1
pywavelets==1.4.1 ; python_version >= '3.8'
requests==2.31.0 ; python_version >= '3.7'
scikit-learn==1.3.2
scipy==1.11.3 ; python_version < '3.13' and python_version >= '3.9'
six==1.16.0 ; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
sympy==1.12 ; python_version >= '3.8'
threadpoolctl==3.2.0 ; python_version >= '3.8'
torch==2.1.0
torchdata==0.7.0 ; python_version >= '3.8'
torchtext==0.16.0
tqdm==4.66.1 ; python_version >= '3.7'
typing-extensions==4.8.0 ; python_version >= '3.8'
tzdata==2023.3 ; python_version >= '2'
urllib3==1.26.18 ; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4, 3.5'
wheel==0.41.3
